move action to selectionclass
